requirements.txt
azure-identity==1.15.0
httpx==0.27.0